refactor(Search): wrap ComposableResult + SourceDefinition + SearchResult under Search; nest PackageQuery.QueryIntent; rename files
Search namespace:
- ComposableResult (12 types: ResultAtom, DocAtom, PackageAtom, ResultSection,
  SourceHint, QuickLink, ComposedSearchResult, ComposedResultBuilder, QueryIntent,
  SourceProperties, SourcePropertiesRegistry, UnifiedSearchSummary) wrapped under
  extension Search.
- SourceDefinition + SourceRegistry wrapped under extension Search.
- FrameworkAvailability + MatchedSymbol wrapped under extension Search.
- Sample.Atom keeps its cross-cutting home; uses SearchModule.ResultAtom to
  disambiguate when Sample shadows Search.
- detectQueryIntent free function retargeted at Search.QueryIntent.

Collision fix:
- Search/PackageQuery.swift had its own QueryIntent enum at Search.QueryIntent,
  conflicting with the broader ComposableResult QueryIntent. Nested it inside
  the actor as Search.PackageQuery.QueryIntent. IntentClassifier + IntentConfig
  references updated to qualify.

In-module sweep:
- Search.Index+Indexing.swift: FrameworkAvailability -> Search.FrameworkAvailability
- Search.Index+Search.swift: MatchedSymbol -> Search.MatchedSymbol;
  SourceRegistry -> Search.SourceRegistry
- Services/UnifiedSearchFormatters.swift: MatchedSymbol -> Search.MatchedSymbol
- Search+SourceDefinition.swift: Search.Source extension qualifies SourceDefinition,
  SourceProperties, SourceRegistry, QueryIntent.

File renames per <Namespace>+<Type>.swift convention:
- ComposableResult.swift     -> Search+ComposableResult.swift
- SourceDefinition.swift     -> Search+SourceDefinition.swift
- SearchResult.swift         -> Search+SearchResult.swift
- PackageQuery.swift         -> Search+PackageQuery.swift

Warning fix:
- ServeReaper.swift: String(cString: [CChar]) deprecation. Added
  nullTerminatedString(from:) helper using String(bytes:encoding:).

Verification: xcrun swift build clean (0 warnings, 0 errors); 1300/1300 tests pass.
